python/requirements.txt
beautifulsoup4==4.9.3
PyAudio==0.2.11
pyttsx3==2.90
selenium==3.141.0
requests-oauthlib==1.3.0
SpeechRecognition==3.8.1
webdriver-manager==3.2.2